requirements.txt
google-genai
openai
fastapi
uvicorn
python-dotenv
